refactor(manager): dedup fence diagnosis; delete dead code; clippy sweep
Readability pass (logic only, behavior-preserving):
- diagnose_post_txn_fence(): the F149 fence-vs-extra_cmp failure
  diagnosis (leader-key read -> displaced/leader flag flips ->
  NotLeader) existed byte-near-identical in txn_fenced and
  txn_fenced_put_revision; now one helper, callers keep their one-line
  fall-through. empty_extra_cmp_err() replaces 3 copies of the
  anomaly error.
- Dead code (workspace-verified zero callers): deleted
  build_extent_inflight_release_op + its discarded del_op/let _ dance
  in drain_extent_inflight_marker; inflight_snapshot_three (stale doc
  claimed a merge-handler caller); LeaseRegistry snapshot/snapshot_for;
  PartitionMetricsWindow::push_with_cap.
- inode_lease.rs: moved a stray section comment out of the middle of
  drain_invalidations's doc comment; clippy --fix (2 first().is_some(),
  1 same-type cast).

Verified: manager lib 169/169; etcd-backed system_split_ref_counting
green (drives split txn + mirrors + owner locks through the refactored
fence path); workspace check green; manager clippy-clean.

+++ crates/manager/src/rpc_handlers.rs
@@ -2032,7 +2032,7 @@ impl AutumnManager {
         };
         if !already_sealed {
             tail.sealed = true;
-            tail.sealed_length = sealed_len as u64;
+            tail.sealed_length = sealed_len;
             tail.eversion += 1;
             tail.avali = avali;
         }
